+++ src/common/helper/PathHelper.h
@@ -30,15 +30,16 @@ class PathHelper
 {
 public:
     static bool isPathValid(const QString &path);
-    static QString extractPath(const QString &path);
+    static QString extractParentDirectory(const QString &path);
     static QString extractFilename(const QString &path);
+    static QString extractFilenameWithFormat(const QString &path);
     static QString extractFormat(const QString &path);
     static QString replaceDateTimeWildcards(const QString &filename);
-    static QString makeUniqueFilename(const QString &path, const QString &filename, const QString &extension = QString());
+    static QString makeUniqueFilename(const QString &path, const QString &filename, const QString &format = QString());
 	static QString replaceNumberWildCards(const QString &filename, const QString &directory, const QString &format);
 
 private:
-	static int getHighestWildcardNumber(const QString &directory, const QString &format, const QString &leftPart, const QString &rightPartWithFormat);
+	static int getHighestWildcardNumber(const QString &directory, const QString &leftPart, const QString &rightPart, const QString &format);
 };
 
 #endif // KSNIP_PATHHELPER_H